admin/requirements.txt
streamlit
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
sentence-transformers
boto3
python-dotenv
pypdf
